chore: move inpure dependencies to separate directory for auth module [ci:force]
