+++ lib/core/api-factories/angular-api-client.factory.ts
@@ -1,6 +1,5 @@
-import { ApiClientFactory, Constructor } from '@alfresco/adf-core/api';
+import { ApiClientFactory, Constructor, AlfrescoApiV2 } from '@alfresco/adf-core/api';
 import { Injectable } from '@angular/core';
-import { AlfrescoApiV2 } from '../api/alfresco-api-v2';
 
 @Injectable()
 export class AngularClientFactory implements ApiClientFactory {
