+++ lib/core/api/index.ts
@@ -1,2 +1,1 @@
 export * from './public-api';
-
