+++ lib/core/api/public-api.ts
@@ -1,4 +1,4 @@
-export * from './alfresco-api-v2-loader.service';
+export * from './alfresco-api-v2';
 export * from './api-client.factory';
 export * from './api-clients.service';
 export * from './clients/index';
