+++ lib/core/auth/src/oidc/auth.module.ts
@@ -1,12 +1,8 @@
 import { APP_INITIALIZER, ModuleWithProviders, NgModule } from '@angular/core';
-import { OAuthModule, OAuthStorage } from 'angular-oauth2-oidc';
-import { StorageService } from '../../../services/storage.service';
-import { BaseAuthenticationService } from '../base-authentication.service';
-import { authConfigFactory, AuthConfigService } from './auth-config.service';
+import { OAuthModule } from 'angular-oauth2-oidc';
 import { AuthRoutingModule } from './auth-routing.module';
-import { AuthModuleConfig, AUTH_CONFIG, AUTH_MODULE_CONFIG } from './auth.module.token';
+import { AuthModuleConfig, AUTH_MODULE_CONFIG } from './auth.module.token';
 import { AuthService } from './auth.service';
-import { OIDCAuthenticationService } from './oidc-authentication.service';
 import { RedirectAuthService } from './redirect-auth.service';
 import { AuthenticationConfirmationComponent } from './view/authentication-confirmation/authentication-confirmation.component';
 
@@ -23,14 +19,7 @@ const defaultConfig: AuthModuleConfig = {
     OAuthModule.forRoot()
   ],
   providers: [
-    {
-        provide: AUTH_CONFIG,
-        useFactory: authConfigFactory,
-        deps: [AuthConfigService]
-    },
-    { provide: BaseAuthenticationService, useClass: OIDCAuthenticationService },
     { provide: AuthService, useExisting: RedirectAuthService },
-    { provide: OAuthStorage, useExisting: StorageService },
     {
       provide: APP_INITIALIZER,
       useFactory: loginFactory,
