+++ lib/core/auth/src/oidc/index.ts
@@ -1,8 +1,9 @@
 export * from './auth-bearer.interceptor';
 export * from './auth-routing.module';
-export * from './oidc-auth.guard';
 export * from './auth.module';
+export * from './auth.module.token';
 export * from './auth.service';
-export * from './oidc-authentication.service';
+export * from './oidc-auth.guard';
 export * from './redirect-auth.service';
 export * from './view/authentication-confirmation/authentication-confirmation.component';
+
